requirements.txt
langchain
langchain-groq
python-dotenv
requests
scikit-learn
numpy
rich
